Changed the class name to represent the library name

+++ common/src/main/java/io/tofpu/multiworldedit/VectorWrapper.java
@@ -4,8 +4,8 @@ public interface VectorWrapper {
     double getX();
     double getY();
     double getZ();
-    double toYaw();
-    double toPitch();
+    float toYaw();
+    float toPitch();
 
     VectorWrapper add(double x, double y, double z);
     VectorWrapper add(final VectorWrapper vectorWrapper);
